requirements.txt
absl-py==2.2.2
alembic==1.15.2
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
cmdstanpy==1.2.5
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cookiecutter==2.6.0
cookiecutter-data-science==2.2.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.71.0
h5py==3.13.0
holidays==0.72
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
loguru==0.7.3
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
nest-asyncio==1.6.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
optuna==4.3.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
prophet==1.1.6
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
ruff==0.11.9
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.7.1
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.7.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.2
